Jira project key can have digits (line 45)
A digit can also be part of the project key for example T3ST is a valid project key and T3ST-123 is thus a valid Jira issue number but it will not be recognized by this tool. 
I propose to add an OR IsDigit() check like: 'while (current.IsAlphaUpper() || current.IsDigit())'
Thanks for considering to upgrade.